+++ collector-media3/src/androidTest/java/com/bitmovin/analytics/media3/Media3ExoPlayerConstants.kt
@@ -1,0 +1,8 @@
+package com.bitmovin.analytics.media3
+
+import com.bitmovin.analytics.systemtest.utils.PlayerInfo
+
+object Media3ExoPlayerConstants {
+    // TODO: what naming should we use??
+    val playerInfo = PlayerInfo("media3-exoplayer", "media3-exoplayer-1.1.1", "Android:Media3")
+}