+++ collector-media3/src/main/java/com/bitmovin/analytics/media3/Media3PlayerCollector.kt
@@ -1,0 +1,59 @@
+package com.bitmovin.analytics.media3
+
+import android.content.Context
+import android.os.Handler
+import androidx.media3.common.util.UnstableApi
+import androidx.media3.exoplayer.ExoPlayer
+import com.bitmovin.analytics.BitmovinAnalytics
+import com.bitmovin.analytics.DefaultCollector
+import com.bitmovin.analytics.adapters.PlayerAdapter
+import com.bitmovin.analytics.api.AnalyticsConfig
+import com.bitmovin.analytics.api.SourceMetadata
+import com.bitmovin.analytics.data.DeviceInformationProvider
+import com.bitmovin.analytics.data.EventDataFactory
+import com.bitmovin.analytics.features.FeatureFactory
+import com.bitmovin.analytics.media3.api.IMedia3ExoPlayerCollector
+import com.bitmovin.analytics.media3.features.Media3ExoPlayerFeatureFactory
+import com.bitmovin.analytics.media3.player.Media3ExoPlayerContext
+import com.bitmovin.analytics.stateMachines.PlayerStateMachine
+import com.bitmovin.analytics.utils.SystemInformationProvider
+import com.bitmovin.analytics.utils.UserAgentProvider
+import com.bitmovin.analytics.utils.Util
+
+@UnstableApi
+internal class Media3PlayerCollector(analyticsConfig: AnalyticsConfig, context: Context) :
+    DefaultCollector<ExoPlayer>(analyticsConfig, context.applicationContext),
+    IMedia3ExoPlayerCollector {
+
+    override var sourceMetadata: SourceMetadata
+        get() = metadataProvider.getSourceMetadata() ?: SourceMetadata()
+        set(value) {
+            metadataProvider.setSourceMetadata(value)
+        }
+
+    override fun createAdapter(
+        player: ExoPlayer,
+        analytics: BitmovinAnalytics,
+    ): PlayerAdapter {
+        val featureFactory: FeatureFactory = Media3ExoPlayerFeatureFactory(analytics, player)
+        val userAgentProvider = UserAgentProvider(
+            Util.getApplicationInfoOrNull(analytics.context),
+            Util.getPackageInfoOrNull(analytics.context),
+            SystemInformationProvider.getProperty("http.agent"),
+        )
+        val eventDataFactory = EventDataFactory(config, userIdProvider, userAgentProvider)
+        val deviceInformationProvider = DeviceInformationProvider(analytics.context)
+        val playerContext = Media3ExoPlayerContext(player)
+        val handler = Handler(player.applicationLooper)
+        val stateMachine = PlayerStateMachine.Factory.create(analytics, playerContext, handler)
+        return Media3ExoPlayerAdapter(
+            player,
+            config,
+            stateMachine,
+            featureFactory,
+            eventDataFactory,
+            deviceInformationProvider,
+            metadataProvider,
+        )
+    }
+}