+++ collector-media3/src/main/java/com/bitmovin/analytics/media3/manipulators/QualityEventDataManipulator.kt
@@ -1,13 +1,15 @@
-package com.bitmovin.analytics.exoplayer.manipulators
+package com.bitmovin.analytics.media3.manipulators
 
+import androidx.media3.common.C.TRACK_TYPE_AUDIO
+import androidx.media3.common.C.TRACK_TYPE_VIDEO
+import androidx.media3.common.Format
+import androidx.media3.common.util.UnstableApi
+import androidx.media3.exoplayer.ExoPlayer
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.manipulators.EventDataManipulator
-import com.bitmovin.analytics.exoplayer.ExoUtil
-import com.google.android.exoplayer2.C.TRACK_TYPE_AUDIO
-import com.google.android.exoplayer2.C.TRACK_TYPE_VIDEO
-import com.google.android.exoplayer2.ExoPlayer
-import com.google.android.exoplayer2.Format
+import com.bitmovin.analytics.media3.Media3Util
 
+@UnstableApi
 internal class QualityEventDataManipulator(private val exoplayer: ExoPlayer) : EventDataManipulator {
     var currentAudioFormat: Format? = null
     var currentVideoFormat: Format? = null
@@ -60,7 +62,7 @@ internal class QualityEventDataManipulator(private val exoplayer: ExoPlayer) : E
 
     fun setFormatsFromPlayer() {
         // TODO: clarify why we are using the fallback here.
-        currentVideoFormat = exoplayer.videoFormat ?: ExoUtil.getSelectedFormatFromPlayer(exoplayer, TRACK_TYPE_VIDEO)
-        currentAudioFormat = exoplayer.audioFormat ?: ExoUtil.getSelectedFormatFromPlayer(exoplayer, TRACK_TYPE_AUDIO)
+        currentVideoFormat = exoplayer.videoFormat ?: Media3Util.getSelectedFormatFromPlayer(exoplayer, TRACK_TYPE_VIDEO)
+        currentAudioFormat = exoplayer.audioFormat ?: Media3Util.getSelectedFormatFromPlayer(exoplayer, TRACK_TYPE_AUDIO)
     }
 }