+++ collector/src/main/java/com/bitmovin/analytics/enums/PlayerType.kt
@@ -4,7 +4,13 @@ import android.os.Parcel
 import android.os.Parcelable
 
 enum class PlayerType(private val value: String) : Parcelable {
-    BITMOVIN("bitmovin"), EXOPLAYER("exoplayer"), AMAZON_IVS("amazonivs");
+    BITMOVIN("bitmovin"),
+    EXOPLAYER("exoplayer"),
+    AMAZON_IVS("amazonivs"),
+
+    // TODO: decide on naming
+    MEDIA3_EXOPLAYER("media3-exoplayer"),
+    ;
 
     override fun toString(): String {
         return value